src/requirements.txt
pytest==6.2.3
Flask==1.1.2
flask-restful==0.3.9
flask_cors==3.0.10
numpy==1.20.2
requests==2.25.1
python-slugify==5.0.2
Pillow==8.2.0